Load Devices Tests are working.
