+++ pytest.ini
@@ -1,2 +1,6 @@
 [pytest]
 pythonpath = tests/pydreo custom_components/dreo/pydreo
+log_cli = 1
+log_cli_level = DEBUG
+log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
+log_cli_date_format=%Y-%m-%d %H:%M:%S
